test: add comprehensive test suite in tests_new/ for all implementations
Cover all 7 implementations: emhash5/6/7/8 and emilib2ss/2o/2s.

verify/ - Quick validation tests:
- test_all_maps.cpp: unified verification of all 7 maps
- quick_test8.cpp / test_emhash58.cpp / test_hidden_bugs8.cpp: emhash8
  focused coverage
- test_extreme.cpp / test_interface_combo.cpp: edge case combinations
- test_emhash5_verify.cpp / test_emhash_allocator.cpp / test_functional_edge.cpp

stress/ - Stress and crash tests:
- stress_all_maps.cpp: 7 maps x 5 items x 1000 trials
- stress_fix.cpp / test_emhash5_stress.cpp / test_emhash5_hifi.cpp
- highload_test.cpp: high load factor scenarios

attack/ - Hash attack tests (require -DEMH_SAFE_PSL):
- hash_attack_all.cpp: unified attack tests
- hash_attack.cpp / hash_attack7.cpp: version-specific attacks

debug/ - Debug helpers (built with -g -O0):
- debug_all_maps.cpp: unified debug suite
- debug_chain.cpp / debug_set_erase.cpp / min_repro.cpp
- reproduce_crash.cpp / fuzz_reproduce.cpp

fuzz/ - Fuzz tests (clang+libfuzzer or ASan):
- fuzz_emhash_all.cpp / fuzz_emilib_all.cpp: unified fuzzers
- fuzz_emhash8.cpp / fuzz_emhash8_advanced.cpp: emhash8 fuzzers
- fuzz_extreme.cpp / fuzz_nocoll.cpp: non-fuzzer stress
- reproduce_emhash8_bug.cpp: bug reproducer

bench/ - Performance benchmarks:
- ebench.cpp / martin_bench.cpp / highload_bench.cpp

+++ tests_new/debug/debug_set_erase.cpp
@@ -136,7 +136,7 @@ int main(int argc, char** argv) {
                 break;
             }
             case OP_COUNT: {
-                fprintf(stderr, "  COUNT key=%d em=%zu ref=%zu\n", op.key, em.count(op.key), ref.count(op.key));
+                fprintf(stderr, "  COUNT key=%d em=%zu ref=%zu\n", op.key, (size_t)em.count(op.key), ref.count(op.key));
                 break;
             }
             case OP_CONTAINS: {
